It was too hard to translate stage coords into coords of the screen
